requirements.txt
aiodocker==0.21.0
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.1.0
black==22.12.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
Deprecated==1.2.13
frozenlist==1.3.3
gitdb==4.0.10
GitPython==3.1.29
idna==3.4
multidict==6.0.3
mypy-extensions==0.4.3
packaging==22.0
pathspec==0.10.3
platformdirs==2.6.0
pycparser==2.21
PyGithub==1.57
PyJWT==2.6.0
PyNaCl==1.5.0
python-dotenv==0.21.0
requests==2.28.1
smmap==5.0.0
tqdm==4.64.1
typing_extensions==4.4.0
urllib3==1.26.13
websocket-client==1.4.2
wrapt==1.14.1
yarl==1.8.2
